% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_xml}
\alias{check_xml}
\title{Validate the XML file}
\usage{
check_xml(name)
}
\arguments{
\item{name}{A string containing the name of the file to load}
}
\value{
For XML files, returns nothing. For non-XML files, returns an error.
The error will get captured by the reactive server_xml_to_dataframe in app.R
and displayed to the user by Shiny.
}
\description{
Validate the file is an XML file by check the file extension
Validate the XML file has the expected structure: \if{html}{\out{<Item>}} and \if{html}{\out{<Field>}} tags
with \if{html}{\out{<Field>}} having the attribute "Name"
The following Field Names are present: 'Genre', 'Composer', 'Work',
'Orchestra', 'Year Recorded', 'Album', 'Conductor', 'Soloists'
}
\examples{
check_xml('test/library.xml')
}
