% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{expand_df}
\alias{expand_df}
\title{Expand dataframe to include 'Orchestra or Soloist' field}
\usage{
expand_df(tag_df)
}
\arguments{
\item{tag_df}{a dataframe generated by the 'filter_dataframe' function}
}
\value{
a dataframe with a new field 'Orchestra or Soloist'
}
\description{
This function creates a new field 'Orchestra or Soloist' which is 1 if either
the 'Orchestra' or 'Soloists' fields have a value, and NA otherwise.
}
\details{
Intended to be used with completeness functions
}
\examples{
columns_to_retain = c('Genre', 'Composer', 'Work', 'Orchestra', 'Year Recorded', 'Album', 'Conductor', 'Soloists')
tag_df = xml_to_dataframe(system.file('extdata', 'library.xml', package = 'jRiverExplorer'))
tag_df = filter_dataframe(tag_df, columns_to_retain)
tag_df = expand_df(tag_df)
}
