% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{filter_dataframe}
\alias{filter_dataframe}
\title{Filter a dataframe of field (tag) values}
\usage{
filter_dataframe(
  dataframe,
  columns_to_retain = c("Genre", "Composer", "Work", "Orchestra", "Year Recorded",
    "Album", "Conductor", "Soloists")
)
}
\arguments{
\item{dataframe}{a dataframe generated from an XML file by the function
xml_to_dataframe}

\item{columns_to_retain}{a character vector of column names to retain}
}
\value{
a dataframe containing only the columns specified in the
columns_to_retain argument. Filtered for unique rows.
}
\description{
This function filters a dataframe to retain only the columns specified in
the columns_to_retain argument. It then removes duplicate rows.
}
\examples{
columns_to_retain = c('Genre', 'Composer', 'Work', 'Orchestra', 'Year Recorded', 'Album', 'Conductor', 'Soloists')
tag_df = xml_to_dataframe(system.file('extdata', 'library.xml', package = 'jRiverExplorer'))
tag_df = filter_dataframe(tag_df, columns_to_retain)
}
