% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{jRiverExplorer}
\alias{jRiverExplorer}
\title{jRiverExplorer
Wrapper function for Shiny app to explore JRiver Media Center Library}
\usage{
jRiverExplorer()
}
\description{
jRiverExplorer
Wrapper function for Shiny app to explore JRiver Media Center Library
}
