% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{make_long_tag_df}
\alias{make_long_tag_df}
\title{Convert the tag_df to long format}
\usage{
make_long_tag_df(tag_df)
}
\arguments{
\item{tag_df}{a dataframe generated by the 'filter_dataframe' function}
}
\value{
a dataframe in long format, with NA values removed
}
\description{
This function takes a dataframe from 'filter_dataframe' or make_tag_completeness_df
and converts it to long format. It then filters out any rows with NA values.
}
\examples{
columns_to_retain = c('Genre', 'Composer', 'Work', 'Orchestra', 'Year Recorded', 'Album', 'Conductor', 'Soloists')
tag_df = xml_to_dataframe(system.file('extdata', 'library.xml', package = 'jRiverExplorer'))
tag_df = filter_dataframe(tag_df, columns_to_retain)
tag_df = make_long_tag_df(tag_df)
}
