% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{xml_to_dataframe}
\alias{xml_to_dataframe}
\title{Convert XML file to dataframe}
\usage{
xml_to_dataframe(file)
}
\arguments{
\item{file}{A string containing the name of the file to load}
}
\value{
A dataframe containing the fields and values for each item. The
dataframe will have one row per item. Columns will be the field names and
values will be the field values.
}
\description{
Convert XML file to dataframe
}
\examples{
file = system.file('extdata', 'library.xml', package = 'jRiverExplorer')
xml_to_dataframe(file)
}
